requirements.txt
fastapi[standard]
imblearn
joblib
matplotlib
numpy
pandas
pyyaml
scikit-learn
scipy
seaborn
streamlit
tqdm
xgboost
uvicorn